Drop install, dev, and CI mode
smuggle does not install anything under the network model, it hijacks. The
install command was a wrapper that started the proxy and blocked while you ran
your own package manager, so `hijack` says what it does and bare `smuggle`
runs it. Package selection moves into session.rs alongside the proxy it feeds.

dev was dead weight: a running dev server resolves from node_modules and never
makes a registry request, so repacking and restarting it changed nothing that
the server could observe.

CI mode was built around install-once semantics that cannot exist when
interception lasts exactly as long as the process. Removed along with the
NDJSON events and job summary; to be redone against the new model.

pm.rs and filetime went with them, having existed only for package manager
detection and bundler cache invalidation on the install path.
